requirements.txt
aiohttp==3.6.3
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
discord.py==1.5.1
idna==2.10
multidict==4.7.6
oyaml==1.0
python-dotenv==0.15.0
PyYAML==5.3.1
yarl==1.5.1
